Move the caddy package into the server package
